moved js tests, added stub test files so that coverage reports are more accurate
