+++ dokomoforms/static/src/survey/js/api/__tests__/FacilityAPI-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('FacilityAPI', () => {
+    var FacilityAPI;
+
+    beforeEach(function() {
+        jest.dontMock('../FacilityAPI.js');
+        FacilityAPI = require('../FacilityAPI');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
