+++ dokomoforms/static/src/survey/js/api/__tests__/PhotoAPI-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('PhotoAPI', () => {
+    var PhotoAPI;
+
+    beforeEach(function() {
+        jest.dontMock('../PhotoAPI.js');
+        PhotoAPI = require('../PhotoAPI');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
