+++ dokomoforms/static/src/survey/js/components/__tests__/Facility-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('Facility', () => {
+    var Facility;
+
+    beforeEach(function() {
+        jest.dontMock('../Facility.js');
+        Facility = require('../Facility');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
