+++ dokomoforms/static/src/survey/js/components/__tests__/Loading-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('Loading', () => {
+    var Loading;
+
+    beforeEach(function() {
+        jest.dontMock('../Loading.js');
+        Loading = require('../Loading');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
