+++ dokomoforms/static/src/survey/js/components/__tests__/Photo-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('Photo', () => {
+    var Photo;
+
+    beforeEach(function() {
+        jest.dontMock('../Photo.js');
+        Photo = require('../Photo');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
