+++ dokomoforms/static/src/survey/js/components/__tests__/Splash-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('Splash', () => {
+    var Splash;
+
+    beforeEach(function() {
+        jest.dontMock('../Splash.js');
+        Splash = require('../Splash');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
