+++ dokomoforms/static/src/survey/js/components/__tests__/Submit-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('Submit', () => {
+    var Submit;
+
+    beforeEach(function() {
+        jest.dontMock('../Submit.js');
+        Submit = require('../Submit');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
