+++ dokomoforms/static/src/survey/js/services/__tests__/auth-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('auth', () => {
+    var auth;
+
+    beforeEach(function() {
+        jest.dontMock('../auth.js');
+        auth = require('../auth');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
