+++ dokomoforms/static/src/survey/js/services/__tests__/location-tests.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+
+// jest.autoMockOff();
+
+// a noop function useful for passing into components that require it.
+var noop = () => {};
+
+describe('location', () => {
+    var location;
+
+    beforeEach(function() {
+        jest.dontMock('../location.js');
+        location = require('../location');
+    });
+
+    it('does nothing', () => {
+
+    });
+
+});
