+++ package.json
@@ -92,41 +92,37 @@
       "js"
     ],
     "unmockedModulePathPatterns": [
-      "<rootDir>/node_modules/react",
-      "<rootDir>/node_modules/react-dom",
-      "<rootDir>/node_modules/react-addons-test-utils",
-      "<rootDir>/node_modules/fbjs"
+      "<rootDir>/node_modules/"
     ],
-    "testDirectoryName": "tests/js",
     "collectCoverage": true,
     "collectCoverageOnlyFrom": {
-      "dokomoforms/static/src/survey/js/Application.js": false,
-      "dokomoforms/static/src/survey/js/api/FacilityAPI.js": false,
-      "dokomoforms/static/src/survey/js/api/PhotoAPI.js": false,
-      "dokomoforms/static/src/common/js/persona.js": false,
-      "dokomoforms/static/src/survey/js/components/Facility.js": false,
-      "dokomoforms/static/src/survey/js/components/Footer.js": false,
-      "dokomoforms/static/src/survey/js/components/Header.js": false,
-      "dokomoforms/static/src/survey/js/components/Location.js": false,
-      "dokomoforms/static/src/survey/js/components/MultipleChoice.js": false,
-      "dokomoforms/static/src/survey/js/components/Note.js": false,
-      "dokomoforms/static/src/survey/js/components/Photo.js": false,
-      "dokomoforms/static/src/survey/js/components/Question.js": false,
-      "dokomoforms/static/src/survey/js/components/Splash.js": false,
-      "dokomoforms/static/src/survey/js/components/Submit.js": false,
+      "dokomoforms/static/src/survey/js/Application.js": true,
+      "dokomoforms/static/src/survey/js/api/FacilityAPI.js": true,
+      "dokomoforms/static/src/survey/js/api/PhotoAPI.js": true,
+      "dokomoforms/static/src/common/js/persona.js": true,
+      "dokomoforms/static/src/survey/js/components/Facility.js": true,
+      "dokomoforms/static/src/survey/js/components/Footer.js": true,
+      "dokomoforms/static/src/survey/js/components/Header.js": true,
+      "dokomoforms/static/src/survey/js/components/Location.js": true,
+      "dokomoforms/static/src/survey/js/components/MultipleChoice.js": true,
+      "dokomoforms/static/src/survey/js/components/Note.js": true,
+      "dokomoforms/static/src/survey/js/components/Photo.js": true,
+      "dokomoforms/static/src/survey/js/components/Question.js": true,
+      "dokomoforms/static/src/survey/js/components/Splash.js": true,
+      "dokomoforms/static/src/survey/js/components/Submit.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/BigButton.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/LittleButton.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/Card.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/Message.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/DontKnow.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/Menu.js": true,
-      "dokomoforms/static/src/survey/js/components/baseComponents/FacilityRadios.js": false,
-      "dokomoforms/static/src/survey/js/components/baseComponents/PhotoField.js": false,
+      "dokomoforms/static/src/survey/js/components/baseComponents/FacilityRadios.js": true,
+      "dokomoforms/static/src/survey/js/components/baseComponents/PhotoField.js": true,
       "dokomoforms/static/src/survey/js/components/baseComponents/PhotoPreview.js": true,
-      "dokomoforms/static/src/survey/js/components/baseComponents/ResponseField.js": false,
-      "dokomoforms/static/src/survey/js/components/baseComponents/ResponseFields.js": false,
-      "dokomoforms/static/src/survey/js/components/baseComponents/Select.js": false,
-      "dokomoforms/static/src/survey/js/components/baseComponents/Title.js": false
+      "dokomoforms/static/src/survey/js/components/baseComponents/ResponseField.js": true,
+      "dokomoforms/static/src/survey/js/components/baseComponents/ResponseFields.js": true,
+      "dokomoforms/static/src/survey/js/components/baseComponents/Select.js": true,
+      "dokomoforms/static/src/survey/js/components/baseComponents/Title.js": true
     }
   },
   "scripts": {
